requirements.txt
aiocontextvars==0.2.2
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
contextvars==2.4
cryptography==36.0.1
decorator==5.1.1
Django==3.2.12
django-cors-headers==3.10.1
django-filter==21.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
hvac==0.11.2
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
loguru==0.6.0
Markdown==3.3.6
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
validators==0.18.2
win32-setctime==1.1.0
zipp==3.6.0
